feat: 주문 결제 기능 연동

+++ src/main/java/com/example/pdelivery/order/application/provider/OrderInfo.java
@@ -12,6 +12,6 @@ public record OrderInfo(UUID orderId,
 						String address,
 						OrderStatus status,
 						String reason,
-						Integer totalPrice,
+						Long totalPrice,
 						List<OrderLineVO> orderLines) {
 }
